+++ client/src/contexts/index.tsx
@@ -1,6 +1,5 @@
 import React, { FC } from "react";
 
-import AuthenticationContextProvider from "./authentication-context/context-provider";
 import CartContextProvider from "./cart-context/context-provider";
 import UserContextProvider from "./user-context/context-provider";
 
@@ -9,11 +8,9 @@ interface IProps {
 }
 
 const ApplicationContextProvider: FC<IProps> = ({ children }) => (
-  <AuthenticationContextProvider>
-    <UserContextProvider>
-      <CartContextProvider>{children}</CartContextProvider>
-    </UserContextProvider>
-  </AuthenticationContextProvider>
+  <UserContextProvider>
+    <CartContextProvider>{children}</CartContextProvider>
+  </UserContextProvider>
 );
 
 export default ApplicationContextProvider;
