--- client/src/contexts/login-userContext/context.js
@@ -1,12 +1,0 @@
-// import React from "react";
-
-// export default React.createContext({
-//   isAuthenticated: false,
-//   setisAuthenticated: () => {},
-//   user: undefined,
-//   setUser: () => {},
-//   login: () => {
-//     console.log("i user context");
-//   },
-//   logout: () => {},
-// });
